# Conflicts:
#	my/messages/en/superadmin.php

+++ sommerce/modules/admin/controllers/traits/settings/PagesTrait.php
@@ -9,6 +9,7 @@ use sommerce\controllers\CommonController;
 use sommerce\helpers\UiHelper;
 use sommerce\modules\admin\components\Url;
 use sommerce\modules\admin\models\forms\EditPageForm;
+use sommerce\modules\admin\models\forms\SavePageForm;
 use sommerce\modules\admin\models\search\PagesSearch;
 use sommerce\modules\admin\models\search\UrlsSearch;
 use Yii;
@@ -48,7 +49,7 @@ trait PagesTrait {
     {
         $this->view->title = Yii::t('admin', "settings.pages_create_page");
 
-        $pageForm = new EditPageForm();
+        $pageForm = new SavePageForm();
         $pageForm->setUser(Yii::$app->user);
         $pageForm->setPage(new Pages());
 
@@ -65,7 +66,7 @@ trait PagesTrait {
             'pageForm' => $pageForm,
             'isNewPage' => $pageForm->getPage()->isNewRecord,
             'storeUrl' => Yii::$app->store->getInstance()->getBaseSite(),
-            'actionUrl' => Url::toRoute('/settings/edit-page'),
+            'actionUrl' => Url::toRoute('/settings/new-page'),
         ]);
     }
 
@@ -133,6 +134,40 @@ trait PagesTrait {
         ];
     }
 
+
+    /**
+     * Create page AJAX action
+     * @param $id
+     * @return array
+     * @throws NotFoundHttpException
+     */
+    public function actionNewPage($id = null)
+    {
+        $request = Yii::$app->getRequest();
+        $response = Yii::$app->getResponse();
+        $response->format = Response::FORMAT_JSON;
+
+        if (!$request->isAjax) {
+            exit;
+        }
+
+        $pageForm = new SavePageForm();
+        $pageForm->setUser(Yii::$app->user);
+
+        if (!$pageForm->edit($request->post(), $id)) {
+            return [
+                'success' => false,
+                'message' => ActiveForm::firstError($pageForm, true)
+            ];
+        };
+
+        return [
+            'success' => true,
+            'message' => Yii::t('admin', 'settings.pages_message_updated'),
+            'id' => $pageForm->getPage()->id,
+        ];
+    }
+
     /**
      * Virtual deleting `page`
      * @param $id
@@ -153,6 +188,10 @@ trait PagesTrait {
         $pageModel = Pages::findOne($id);
         if (!$pageModel) {
             throw new NotFoundHttpException();
+        }
+
+        if (!Pages::canDelete($pageModel)) {
+            throw new ForbiddenHttpException();
         }
 
         $pageModel->deleteVirtual();
